Basic structure of training process
So far, only testing whether it is possible to perform fitting and
predictions with the default values of the classifier. There is no
need to make it any more complicated before *this* does not work.

Next step will the implementation of a proper grid search process,
although this has to take into account the selected matrix.